clinical-extraction/pyproject.toml
[project]
name = "clinical-extraction"
version = "0.1.0"
description = "Clinical data extraction pipeline: VCF variant + pathology report (PDF) processing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LangGraph / LangChain
    "langgraph>=1.0.7",
    "langchain>=0.3.0",
    "langchain-aws>=1.4.1",
    # AgentCore + DeepAgent
    "bedrock-agentcore>=1.2.0",
    "langchain-agentcore-codeinterpreter>=0.0.1",
    "deepagents>=0.1.0",
    # PDF processing
    "pdfplumber>=0.10.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    # Observability
    "langfuse>=2.0.0",
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Infrastructure
    "boto3>=1.42.37",
    "botocore[crt]>=1.42.76",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "pymupdf>=1.27.2.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "jupyter>=1.1.1",
    "langgraph-cli[inmem]>=0.4.19",
]
